ai/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.1.2
cffi==1.17.1
click==8.1.8
cryptography==46.0.3
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.124.4
greenlet==2.0.2
h11==0.16.0
httptools==0.6.4
idna==3.11
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.4.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
